pyproject.toml
[tool.poetry]
name = "dlsbot"
version = "0.0.114514"
description = "提供校内服务"
authors = ["eya46 <admin@eya46.com>"]
license = "Apache-2.0 license"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "2.0.0"
nonebot-plugin-apscheduler = "^0.3.0"
gino = "^1.0.1"
nonebot-adapter-onebot = "^2.2.4"
httpx = "^0.25.0"
pillow = "^10.0.0"
wcwidth = "^0.2.6"
jinja2 = "^3.1.2"
playwright = "1.36.0"
nonebot-plugin-htmlrender = "^0.2.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
